ref(slack): migrate slack parser out of platforms
